Experimental code for FTC robotics
This stuff should run on a regular IDE, and should be pretty similar to things on the robot.  Where necessary, some stubs have been added.
It's for educational purposes and intended primarily for the benefit of the students involved in FTC teams 8578 and 11959, but it is what it is, and if anyone else can derive some education or use out of it, that's good too.